Move Crisis Support to standalone Wellness tab and not under Resources
Refactor crisis support resources: remove Support component and integrate its content into PhoneNumber component; update references in Resources and Wellness pages.
