+++ client/src/App.tsx
@@ -55,7 +55,6 @@ const Testing = lazy(() => import('./pages/Testing'));
 const NYTimes = lazy(() => import('./components/resources/NYTimes'));
 const Adobe = lazy(() => import('./components/resources/Adobe'));
 const LibraryCard = lazy(() => import('./components/resources/LibraryCard'));
-const Support = lazy(() => import('./components/resources/Support'));
 const PhoneNumber = lazy(() => import('./components/wellness/PhoneNumber'));
 
 
@@ -157,7 +156,6 @@ export default function App() {
                             <Route path="nytimes" element={<NYTimes />} />
                             <Route path="adobe" element={<Adobe />} />
                             <Route path="library-card" element={<LibraryCard />} />
-                            <Route path="support" element={<Support />} />
                         </Route>
                         <Route
                             path="/super-secret-testing"
